+++ webviz_subsurface/_providers/ensemble_surface_provider/ensemble_surface_provider_factory.py
@@ -21,9 +21,15 @@ LOGGER = logging.getLogger(__name__)
 
 
 class EnsembleSurfaceProviderFactory(WebvizFactory):
-    def __init__(self, root_storage_folder: Path, allow_storage_writes: bool) -> None:
+    def __init__(
+        self,
+        root_storage_folder: Path,
+        allow_storage_writes: bool,
+        avoid_copying_surfaces: bool,
+    ) -> None:
         self._storage_dir = Path(root_storage_folder) / __name__
         self._allow_storage_writes = allow_storage_writes
+        self._avoid_copying_surfaces = avoid_copying_surfaces
 
         LOGGER.info(
             f"EnsembleSurfaceProviderFactory init: storage_dir={self._storage_dir}"
@@ -41,8 +47,13 @@ class EnsembleSurfaceProviderFactory(WebvizFactory):
             app_instance_info = WEBVIZ_FACTORY_REGISTRY.app_instance_info
             storage_folder = app_instance_info.storage_folder
             allow_writes = app_instance_info.run_mode != WebvizRunMode.PORTABLE
+            dont_copy_surfs = app_instance_info.run_mode == WebvizRunMode.NON_PORTABLE
 
-            factory = EnsembleSurfaceProviderFactory(storage_folder, allow_writes)
+            factory = EnsembleSurfaceProviderFactory(
+                root_storage_folder=storage_folder,
+                allow_storage_writes=allow_writes,
+                avoid_copying_surfaces=dont_copy_surfs,
+            )
 
             # Store the factory object in the global factory registry
             WEBVIZ_FACTORY_REGISTRY.set_factory(EnsembleSurfaceProviderFactory, factory)
@@ -80,11 +91,14 @@ class EnsembleSurfaceProviderFactory(WebvizFactory):
         obs_surface_files = discover_observed_surface_files(ens_path, attribute_filter)
         et_discover_s = timer.lap_s()
 
+        # As an optimization, avoid copying the surfaces into the backing store,
+        # typically when  we're running in non-portable mode
         ProviderImplFile.write_backing_store(
             self._storage_dir,
             storage_key,
             sim_surfaces=sim_surface_files,
             obs_surfaces=obs_surface_files,
+            avoid_copying_surfaces=self._avoid_copying_surfaces,
         )
         et_write_s = timer.lap_s()